Rename sbStressBase.ts file to match the name of the contained class (ServiceBusStressTester)
