+++ sdk/servicebus/service-bus/test/stress/scenarioLongRunning.ts
@@ -1,4 +1,7 @@
-import { captureConsoleOutputToAppInsights, ServiceBusStressTester } from "./stressTestsBase";
+import {
+  captureConsoleOutputToAppInsights,
+  ServiceBusStressTester
+} from "./serviceBusStressTester";
 import { AbortController, AbortSignalLike } from "@azure/abort-controller";
 import { ServiceBusClient, ServiceBusSender } from "@azure/service-bus";
 import { v4 as uuidv4 } from "uuid";
