+++ sdk/servicebus/service-bus/test/stress/scenarioPeekMessages.ts
@@ -3,7 +3,7 @@ import {
   ServiceBusReceivedMessage,
   ServiceBusReceiver
 } from "@azure/service-bus";
-import { ServiceBusStressTester } from "./stressTestsBase";
+import { ServiceBusStressTester } from "./serviceBusStressTester";
 import { delay } from "rhea-promise";
 import parsedArgs from "minimist";
