+++ sdk/servicebus/service-bus/test/stress/scenarioRenewSessionLock.ts
@@ -1,5 +1,5 @@
 import { ServiceBusClient, ServiceBusSessionReceiver } from "@azure/service-bus";
-import { ServiceBusStressTester } from "./stressTestsBase";
+import { ServiceBusStressTester } from "./serviceBusStressTester";
 import { delay } from "rhea-promise";
 import parsedArgs from "minimist";
