+++ sdk/servicebus/service-bus/test/stress/scenarioStreamingReceive.ts
@@ -3,7 +3,7 @@ import {
   ServiceBusReceiver,
   ServiceBusReceiverOptions
 } from "@azure/service-bus";
-import { ServiceBusStressTester } from "./stressTestsBase";
+import { ServiceBusStressTester } from "./serviceBusStressTester";
 import { delay } from "rhea-promise";
 import parsedArgs from "minimist";
